fix: project page and thumbnails - Fix thumbnails, add dynamic routes, consolidate content config, add sample projects
