--- src/content.config.ts
@@ -1,20 +1,0 @@
-import { glob } from "astro/loaders";
-import { defineCollection, z } from "astro:content";
-
-const blog = defineCollection({
-  // Load Markdown and MDX files in the `src/content/blog/` directory.
-  loader: glob({ base: "./src/content/blog", pattern: "**/*.{md,mdx}" }),
-  // Type-check frontmatter using a schema
-  schema: z.object({
-    title: z.string(),
-    description: z.string(),
-    // Transform string to Date object
-    pubDate: z.coerce.date(),
-    updatedDate: z.coerce.date().optional(),
-    heroImage: z.string().optional(),
-    // Add tags field as an array of strings
-    tags: z.array(z.string()).optional(),
-  }),
-});
-
-export const collections = { blog };
