+++ src/content/config.ts
@@ -1,0 +1,27 @@
+import { defineCollection, z } from "astro:content";
+
+const blog = defineCollection({
+	schema: z.object({
+		title: z.string(),
+		description: z.string(),
+		pubDate: z.coerce.date(),
+		updatedDate: z.coerce.date().optional(),
+		heroImage: z.string().optional(),
+		tags: z.array(z.string()).optional(),
+	}),
+});
+
+const projects = defineCollection({
+	schema: z.object({
+		title: z.string(),
+		description: z.string(),
+		image: z.string().optional(),
+		tags: z.array(z.string()).optional(),
+		link: z.string().optional(),
+	}),
+});
+
+export const collections = {
+	blog,
+	projects,
+};
