+++ src/content/projects/project1.md
@@ -1,0 +1,8 @@
+---
+title: "Project One"
+description: "This is the first sample project."
+image: "/blog-placeholder-1.jpg"
+tags: ["sample", "project"]
+---
+
+This is the content for Project One.
