+++ src/content/projects/project2.md
@@ -1,0 +1,8 @@
+---
+title: "Project Two"
+description: "This is the second sample project."
+image: "/blog-placeholder-2.jpg"
+tags: ["sample", "project"]
+---
+
+This is the content for Project Two.
