+++ src/content/projects/project3.md
@@ -1,0 +1,8 @@
+---
+title: "Project Three"
+description: "This is the third sample project."
+image: "/blog-placeholder-3.jpg"
+tags: ["sample", "project"]
+---
+
+This is the content for Project Three.
