+++ src/pages/about.astro
@@ -28,7 +28,7 @@ import '../styles/tailwind.css';
 							<div class="flex flex-col items-center mb-6">
 								<div class="w-40 h-40 rounded-full overflow-hidden mb-4">
 									<img 
-										src="/mikenavales-xyz/20200711-v1-450w-kyooreas-mike-navales-web-developer.webp" 
+										src="/20200711-v1-450w-kyooreas-mike-navales-web-developer.webp" 
 										alt="Mike Navales - Web Developer"
 										width={160}
 										height={160}
