+++ src/pages/projects.astro
@@ -5,7 +5,9 @@ import Footer from '../components/Footer.astro';
 import { SITE_TITLE, SITE_DESCRIPTION } from '../consts';
 import '../styles/global.css';
 import '../styles/tailwind.css';
-import { projects } from '../data/projects';
+import { getCollection } from 'astro:content';
+
+const projects = await getCollection('projects');
 ---
 
 <!doctype html>
@@ -25,12 +27,12 @@ import { projects } from '../data/projects';
 				<section class="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-6 px-2 justify-items-center">
 					{projects.map((project) => (
 						<div class="bg-gray-900 rounded-lg overflow-hidden shadow-lg transition-all duration-300 hover:transform hover:scale-105 hover:shadow-[0_5px_15px_-5px_rgba(255,255,255,0.2)] hover:-translate-y-2 w-full max-w-sm group">
-							<a href={project.link} class="block">
+							<a href={project.data.link || `/projects/${project.id}`} class="block">
 								<div class="relative h-40 overflow-hidden">
 									<div class="absolute inset-0 bg-gradient-to-t from-gray-900 to-transparent opacity-0 group-hover:opacity-60 transition-opacity duration-300 z-10"></div>
 									<img 
-										src={project.image} 
-										alt={project.title} 
+										src={project.data.image || '/blog-placeholder-1.jpg'} 
+										alt={project.data.title} 
 										class="w-full h-full object-cover transition-transform duration-500 group-hover:scale-110"
 										loading="lazy"
 										width="800"
@@ -40,10 +42,10 @@ import { projects } from '../data/projects';
 									/>
 								</div>
 								<div class="p-4 transition-colors duration-300 group-hover:bg-gray-800">
-									<h2 class="text-lg md:text-xl font-bold mb-2 text-center group-hover:text-blue-400 transition-colors duration-300">{project.title}</h2>
-									<p class="text-gray-300 text-sm md:text-base mb-3 line-clamp-2 group-hover:text-gray-200 transition-colors duration-300">{project.description}</p>
+									<h2 class="text-lg md:text-xl font-bold mb-2 text-center group-hover:text-blue-400 transition-colors duration-300">{project.data.title}</h2>
+									<p class="text-gray-300 text-sm md:text-base mb-3 line-clamp-2 group-hover:text-gray-200 transition-colors duration-300">{project.data.description}</p>
 									<div class="flex flex-wrap gap-1 justify-center">
-										{project.tags.map((tag) => (
+										{project.data.tags?.map((tag) => (
 											<span class="bg-blue-900 text-blue-200 px-2 py-0.5 rounded-full text-xs group-hover:bg-blue-800 transition-colors duration-300">
 												{tag}
 											</span>
@@ -70,4 +72,4 @@ import { projects } from '../data/projects';
 	.bg-gray-900:hover {
 		box-shadow: 0 5px 15px -5px rgba(255, 255, 255, 0.2), 0 0 10px rgba(255, 255, 255, 0.1);
 	}
-</style> 
+</style>
