+++ src/pages/projects/index.astro
@@ -1,0 +1,30 @@
+---
+import { getCollection } from 'astro:content';
+
+const projects = await getCollection('projects');
+---
+
+<!doctype html>
+<html lang="en">
+  <head>
+    <title>Projects | Mike Navales</title>
+  </head>
+  <body class="bg-black text-white">
+    <main class="w-full bg-black">
+      <div class="max-w-[1920px] mx-auto px-2 py-8">
+        <section class="mb-12 px-2 text-center">
+          <h1 class="text-3xl md:text-4xl font-bold mb-4">Projects</h1>
+          <ul class="list-none">
+            {projects.map((project) => (
+              <li key={project.id} class="mb-4">
+                <a href={`/projects/${project.id}`} class="text-blue-400 hover:underline">
+                  {project.data.title}
+                </a>
+              </li>
+            ))}
+          </ul>
+        </section>
+      </div>
+    </main>
+  </body>
+</html>
